Move ScaledImagesCacheService to BlobStorageModule and ImgproxyService to new ImgProxyModule
