+++ demo/api/src/app.module.ts
@@ -11,6 +11,7 @@ import {
     DamModule,
     DependenciesModule,
     FileUploadsModule,
+    ImgProxyModule,
     KubernetesModule,
     PageTreeModule,
     RedirectsModule,
@@ -135,6 +136,10 @@ export class AppModule {
                 }),
                 BlobStorageModule.register({
                     backend: config.blob.storage,
+                    cacheDirectory: `${config.blob.storageDirectoryPrefix}-cache`,
+                }),
+                ImgProxyModule.register({
+                    imgproxyConfig: config.imgproxy,
                 }),
                 DamModule.register({
                     damConfig: {
@@ -143,10 +148,9 @@ export class AppModule {
                         allowedImageSizes: config.dam.allowedImageSizes,
                         allowedAspectRatios: config.dam.allowedImageAspectRatios,
                         filesDirectory: `${config.blob.storageDirectoryPrefix}-files`,
-                        cacheDirectory: `${config.blob.storageDirectoryPrefix}-cache`,
                         maxFileSize: config.dam.uploadsMaxFileSize,
+                        maxSrcResolution: config.imgproxy.maxSrcResolution,
                     },
-                    imgproxyConfig: config.imgproxy,
                     Scope: DamScope,
                     File: DamFile,
                     Folder: DamFolder,
