+++ packages/api/cms-api/src/blob-storage/blob-storage.config.ts
@@ -4,4 +4,5 @@ import { type BlobStorageS3Config } from "./backends/s3/blob-storage-s3.config";
 
 export interface BlobStorageConfig {
     backend: BlobStorageFileConfig | BlobStorageAzureConfig | BlobStorageS3Config;
+    cacheDirectory: string;
 }
