+++ packages/api/cms-api/src/blob-storage/blob-storage.module.ts
@@ -3,6 +3,7 @@ import { DynamicModule, Global, Module } from "@nestjs/common";
 import { BlobStorageBackendService } from "./backends/blob-storage-backend.service";
 import { BlobStorageConfig } from "./blob-storage.config";
 import { BLOB_STORAGE_CONFIG } from "./blob-storage.constants";
+import { ScaledImagesCacheService } from "./cache/scaled-images-cache.service";
 
 @Global()
 @Module({})
@@ -15,8 +16,8 @@ export class BlobStorageModule {
 
         return {
             module: BlobStorageModule,
-            providers: [blobStorageConfigProvider, BlobStorageBackendService],
-            exports: [BlobStorageBackendService, blobStorageConfigProvider],
+            providers: [blobStorageConfigProvider, BlobStorageBackendService, ScaledImagesCacheService],
+            exports: [BlobStorageBackendService, blobStorageConfigProvider, ScaledImagesCacheService],
         };
     }
 }
