+++ packages/api/cms-api/src/blob-storage/cache/dto/file-cache.interface.ts
@@ -1,0 +1,6 @@
+import { type StorageMetaData } from "../../backends/blob-storage-backend.interface";
+
+export interface FileCache {
+    file: NodeJS.ReadableStream;
+    metaData: StorageMetaData;
+}
