--- packages/api/cms-api/src/dam/cache/dto/file-cache.interface.ts
@@ -1,6 +1,0 @@
-import { type StorageMetaData } from "../../../blob-storage/backends/blob-storage-backend.interface";
-
-export interface FileCache {
-    file: NodeJS.ReadableStream;
-    metaData: StorageMetaData;
-}
