+++ packages/api/cms-api/src/dam/dam.constants.ts
@@ -1,3 +1,2 @@
 export const DAM_CONFIG = "dam-config";
-export const IMGPROXY_CONFIG = "imgproxy-config";
 export const DAM_FILE_VALIDATION_SERVICE = "dam-file-validation-service";
