+++ packages/api/cms-api/src/file-uploads/file-uploads-download.controller.ts
@@ -19,13 +19,13 @@ import { PassThrough, Readable } from "stream";
 
 import { DisableCometGuards } from "../auth/decorators/disable-comet-guards.decorator";
 import { BlobStorageBackendService } from "../blob-storage/backends/blob-storage-backend.service";
+import { ScaledImagesCacheService } from "../blob-storage/cache/scaled-images-cache.service";
 import { createHashedPath } from "../blob-storage/utils/create-hashed-path.util";
-import { ScaledImagesCacheService } from "../dam/cache/scaled-images-cache.service";
 import { calculatePartialRanges } from "../dam/files/files.utils";
 import { ALL_TYPES, BASIC_TYPES, MODERN_TYPES } from "../dam/images/images.constants";
 import { getSupportedMimeType } from "../dam/images/images.util";
-import { Extension, ResizingType } from "../dam/imgproxy/imgproxy.enum";
-import { ImgproxyService } from "../dam/imgproxy/imgproxy.service";
+import { Extension, ResizingType } from "../imgproxy/imgproxy.enum";
+import { ImgproxyService } from "../imgproxy/imgproxy.service";
 import { RequiredPermission } from "../user-permissions/decorators/required-permission.decorator";
 import { DownloadParams, HashDownloadParams, HashImageParams, ImageParams } from "./dto/file-uploads-download.params";
 import { FileUpload } from "./entities/file-upload.entity";
