+++ packages/api/cms-api/src/file-uploads/file-uploads.module.ts
@@ -3,6 +3,7 @@ import { DynamicModule, Global, Module, Provider } from "@nestjs/common";
 
 import { BlobStorageModule } from "../blob-storage/blob-storage.module";
 import { FileValidationService } from "../dam/files/file-validation.service";
+import { ImgProxyModule } from "../imgproxy/imgProxy.module";
 import { FileUpload } from "./entities/file-upload.entity";
 import { FileUploadsConfig } from "./file-uploads.config";
 import { FILE_UPLOADS_CONFIG, FILE_UPLOADS_FILE_VALIDATION_SERVICE } from "./file-uploads.constants";
@@ -48,7 +49,7 @@ export class FileUploadsModule {
 
         return {
             module: FileUploadsModule,
-            imports: [MikroOrmModule.forFeature([FileUpload]), BlobStorageModule],
+            imports: [MikroOrmModule.forFeature([FileUpload]), BlobStorageModule, ImgProxyModule],
             providers,
             controllers,
             exports: [FileUploadsService],
