+++ packages/api/cms-api/src/imgproxy/imgproxy.constants.ts
@@ -1,0 +1,1 @@
+export const IMGPROXY_CONFIG = "imgproxy-config";
