+++ packages/api/cms-api/src/imgproxy/imgproxy.enum.ts
